fix(build): avoid Turbopack chunk collision with GCS path rename
